fix(cc-widgets): shift functions from test file to util file
